commit 1

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,18 +10,18 @@
  */
 int main(void)
 {
-   unsigned long int t0 = 0, t1 = 1, s;
+   long long int t0 = 0, t1 = 1, s;
     int i;
     for (i = 1; i<= 98; i++)
     {
         s = (unsigned long int)t0 + (unsigned long int)t1;
         if (i != 98)
         {
-			printf("%lu, ", s);
+			printf("%lld, ", s);
         }
         else
         {
-            printf("%lu\n", s);
+            printf("%lld\n", s);
         }
         t0 = t1;
         t1 = s;
